--- src/arch/x86_64/registers/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(target_os = "linux")]
-pub mod debug;
